Move Timestamp class from Firestore to Firebase Core (#13221)
