+++ FirebaseCore/Sources/FIRTimestampInternal.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#import "FIRTimestamp.h"
+#import "FirebaseCore/Sources/Public/FirebaseCore/FIRTimestamp.h"
 
 NS_ASSUME_NONNULL_BEGIN
