+++ FirebaseCore/Sources/Public/FirebaseCore/FirebaseCore.h
@@ -18,4 +18,5 @@
 #import "FIRConfiguration.h"
 #import "FIRLoggerLevel.h"
 #import "FIROptions.h"
+#import "FIRTimestamp.h"
 #import "FIRVersion.h"
