+++ Firestore/CHANGELOG.md
@@ -5,6 +5,8 @@
   instructions.
 - [changed] **Breaking change**: LRU Garbage Collector is now the default GC for memory cache, eager GC is now
   opt-in (via MemoryCacheSettings(garbageCollectorSettings: MemoryEagerGCSettings())) instead of the default one.
+- [changed] Move `Timestamp` class into `FirebaseCore`. `FirebaseFirestore.Timestamp`
+  was changed to `FirebaseCore.Timestamp`. (#13221)
 
 # 10.25.0
 - [fixed] Allow blob of data with zero length. (#11773, #12620)
