+++ Firestore/Example/Tests/Integration/API/FIRCompositeIndexQueryTests.mm
@@ -15,6 +15,9 @@
  */
 
 #import <FirebaseFirestore/FirebaseFirestore.h>
+
+#import "FirebaseCore/FIRTimestamp.h"
+
 #import "Firestore/Example/Tests/Util/FSTIntegrationTestCase.h"
 
 #include "Firestore/core/src/util/autoid.h"
