+++ Firestore/Example/Tests/Integration/API/FIRDatabaseTests.mm
@@ -21,6 +21,8 @@
 
 #import <XCTest/XCTest.h>
 
+#import "FirebaseCore/FIRTimestamp.h"
+
 #import "FirebaseCore/Extension/FIRAppInternal.h"
 #import "Firestore/Example/Tests/Util/FSTEventAccumulator.h"
 #import "Firestore/Example/Tests/Util/FSTIntegrationTestCase.h"
