+++ Firestore/Example/Tests/Integration/API/FIRServerTimestampTests.mm
@@ -18,6 +18,8 @@
 
 #import <XCTest/XCTest.h>
 
+#import "FirebaseCore/FIRTimestamp.h"
+
 #import "Firestore/Example/Tests/Util/FSTEventAccumulator.h"
 #import "Firestore/Example/Tests/Util/FSTIntegrationTestCase.h"
 #import "Firestore/Source/API/FIRFirestore+Internal.h"
