+++ Firestore/Source/API/converters.mm
@@ -19,7 +19,7 @@
 #include <utility>
 
 #import "FIRGeoPoint.h"
-#import "FIRTimestamp.h"
+#import "FirebaseCore/FIRTimestamp.h"
 
 #include "Firestore/Source/API/FIRDocumentReference+Internal.h"
 #include "Firestore/core/include/firebase/firestore/geo_point.h"
