+++ Firestore/Swift/Source/Codable/TimestampDecodingStrategy.swift
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-import Foundation
-#if SWIFT_PACKAGE
-  @_exported import FirebaseFirestoreInternalWrapper
-#else
-  @_exported import FirebaseFirestoreInternal
-#endif // SWIFT_PACKAGE
+import FirebaseCore
 import FirebaseSharedSwift
 
 public extension FirebaseDataDecoder.DateDecodingStrategy {
