+++ Firestore/Swift/Tests/API/BasicCompileTests.swift
@@ -20,6 +20,7 @@
 import Foundation
 import XCTest
 
+import FirebaseCore
 import FirebaseFirestore
 
 class BasicCompileTests: XCTestCase {
