+++ Firestore/Swift/Tests/Integration/CodableIntegrationTests.swift
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import FirebaseCore
 @testable import FirebaseFirestore
 import Foundation
